requirements.txt
flask==3.0.3
faker==25.8.0
google-api-python-client==2.160.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
msal==1.28.0
requests==2.32.3
pyyaml==6.0.1
openai==1.65.0
